Sprint #4

See merge request ftmd/25-ftmd/if3250-25-ftmd-sixplus!45

+++ LO/models.py
@@ -49,19 +49,20 @@ class LO(models.Model):
         return lo_dict, course[0]
 
 class Section(models.Model):
-    course_id = models.ForeignKey(Course, on_delete=models.CASCADE)
+    course = models.ForeignKey(Course, on_delete=models.CASCADE)
     sec_id = models.IntegerField()
     semester = models.IntegerField()
     year = models.IntegerField()
 
     def __str__(self):
-        return f"{self.course_id}, K{self.sec_id}, {self.semester}-{self.year}"
+        return f"{self.course}, K{self.sec_id}, {self.semester}-{self.year}"
 
 
 class Takes(models.Model):
     student = models.ForeignKey(Student, on_delete=models.CASCADE)
     section = models.ForeignKey(Section, on_delete=models.CASCADE)
     grade = models.CharField(default='-', max_length=2)
+    isKuesionerFilled = models.BooleanField(default = False)
 
     def __str__(self):
         return f"{self.student}, {self.section}, {self.grade}"
@@ -75,9 +76,11 @@ class Takes(models.Model):
             student_name_list.append(obj.student.name)
         return student_nim_list, student_name_list
 
+    def get_individual_student_takes(self, nim, section):
+        return Takes.objects.filter(student__nim = nim, section = section)
+
 class Score(models.Model):
-    nim = models.ForeignKey(Student, on_delete=models.CASCADE)
-    course = models.ForeignKey(Course, on_delete=models.CASCADE)
+    takes = models.ForeignKey(Takes, on_delete=models.CASCADE)
     uts1 = models.IntegerField()
     uts2 = models.IntegerField()
     uas = models.IntegerField()
@@ -85,19 +88,47 @@ class Score(models.Model):
     tutorial = models.IntegerField()
 
     def __str__(self):
-        return f"{self.nim}, {self.course}, UTS1:{self.uts1}, UTS2:{self.uts2}, UAS:{self.uas}, Kuis:{self.kuis}, Tutorial:{self.tutorial}"
+        return f"{self.takes}, UTS1:{self.uts1}, UTS2:{self.uts2}, UAS:{self.uas}, Kuis:{self.kuis}, Tutorial:{self.tutorial}"
 
-    def getStudentScore(self, nim, course_id):
-        return Score.objects.filter(nim__nim=nim, course__course_id=course_id)
+    def getStudentTakesScores(self, course_id, year, semester, section_id):
+        scores = Score.objects.filter( 
+            takes__section__course__course_id = course_id, 
+            takes__section__year = year, 
+            takes__section__semester = semester,
+            takes__section__sec_id = section_id)
+        score_list = []
 
-    def setStudentScore(self, nim, course_id, nilai_uts1, nilai_uts2, nilai_uas, nilai_kuis, nilai_tutorial):
-        course = Course.objects.filter(course_id=course_id)[0]
-        student = Student.objects.filter(nim=nim)[0]
-        score = self.getStudentScore(self, nim = nim, course_id = course_id)
-        if(len(score) != 0):
-            score[0].delete()
-        new_score = Score.objects.create(nim=student, course=course, uts1=nilai_uts1, uts2=nilai_uts2, uas=nilai_uas, kuis=nilai_kuis, tutorial=nilai_tutorial)
-        return new_score
+        for obj in scores:
+            score_list.append(obj)
+        return scores
+
+    def getStudentScore(self, nim, course_id, year, semester, section_id):
+        return Score.objects.filter(takes__student__nim=nim, 
+            takes__section__course__course_id = course_id, 
+            takes__section__year = year, 
+            takes__section__semester = semester,
+            takes__section__sec_id = section_id)
+
+    def setStudentScore(self, nim, course_id, year, semester, section_id, nilai_uts1, nilai_uts2, nilai_uas, nilai_kuis, nilai_tutorial):
+        course_filter = Course.objects.filter(course_id=course_id)
+        if(len(course_filter) == 0):
+            return False
+        course = course_filter[0]
+
+        student_filter = Student.objects.filter(nim=nim)
+        if(len(student_filter) == 0):
+            return False
+        student = student_filter[0]
+
+        section = Section.objects.get(course__course_id = course_id, sec_id = section_id, year = year, semester = semester)
+        takes = Takes.get_individual_student_takes(Takes, nim, section)
+        if(len(takes) != 0):
+            score = self.getStudentScore(self, nim = nim, course_id = course_id, year = year, semester = semester, section_id = section_id)
+            if(len(score) != 0):
+                score[0].delete()
+            new_score = Score.objects.create(takes = takes[0], uts1=nilai_uts1, uts2=nilai_uts2, uas=nilai_uas, kuis=nilai_kuis, tutorial=nilai_tutorial)
+            return True
+        return False
 
 class BobotKomponenScore(models.Model):
     course = models.ForeignKey(Course, on_delete=models.CASCADE)
@@ -111,8 +142,7 @@ class BobotKomponenScore(models.Model):
         return f"{self.course}, UTS1:{self.uts1}, UTS2:{self.uts2}, UAS:{self.uas}, Kuis:{self.kuis}, Tutorial:{self.tutorial}"
 
 class ResponseKerjasama(models.Model):
-    student = models.ForeignKey(Student, on_delete=models.CASCADE)
-    course = models.ForeignKey(Course, on_delete=models.CASCADE)
+    takes = models.ForeignKey(Takes, on_delete=models.CASCADE)
     Kontribusi = models.IntegerField()
     PemecahanMasalah = models.IntegerField()
     Sikap = models.IntegerField()
@@ -120,12 +150,11 @@ class ResponseKerjasama(models.Model):
     BekerjaDenganOrangLain = models.IntegerField()
 
     def __str__(self):
-        return f"Mahasiswa: {self.student} Matkul: {self.course} Kontribusi: {self.Kontribusi} PemecahanMasalah: {self.PemecahanMasalah} Sikap: {self.Sikap} FokusTerhadapTugas: {self.FokusTerhadapTugas} BekerjaDenganOrangLain: {self.BekerjaDenganOrangLain}"
+        return f"Mahasiswa: {self.takes.student} Matkul: {self.takes.section.course_id} Kontribusi: {self.Kontribusi} PemecahanMasalah: {self.PemecahanMasalah} Sikap: {self.Sikap} FokusTerhadapTugas: {self.FokusTerhadapTugas} BekerjaDenganOrangLain: {self.BekerjaDenganOrangLain}"
         
 class ResponseKomunikasi(models.Model):
-    student = models.ForeignKey(Student, on_delete=models.CASCADE)
+    takes = models.ForeignKey(Takes, on_delete=models.CASCADE)
     kelompok = models.IntegerField(default=1)
-    course = models.ForeignKey(Course, on_delete=models.CASCADE)
     Penyampaian1 = models.IntegerField()
     Penyampaian2 = models.IntegerField()
     Penyampaian3 = models.IntegerField()
@@ -138,4 +167,98 @@ class ResponseKomunikasi(models.Model):
     Waktu = models.IntegerField()
 
     def __str__(self):
-        return f"Mahasiswa: {self.student} Kelompok: {self.kelompok} Matkul: {self.course} CaraPenyampaianInformasi1: {self.Penyampaian1} CaraPenyampaianInformasi2: {self.Penyampaian2} CaraPenyampaianInformasi3: {self.Penyampaian3} CaraPenyampaianInformasi4: {self.Penyampaian4} KontenInformasiYangDisampaikan: {self.Konten} BahasaYangDigunakanDalamPenyampaianInformasi: {self.Bahasa} PenguasaanMateri: {self.Penguasaan} MenjawabPertanyaan: {self.Menjawab} PenggunaanMediaPendukung: {self.Media} MenggunakanWaktuDenganEfektifDanEfisien: {self.Waktu}"
+        return f"Mahasiswa: {self.takes.student} Kelompok: {self.kelompok} Matkul: {self.takes.section.course_id} CaraPenyampaianInformasi1: {self.Penyampaian1} CaraPenyampaianInformasi2: {self.Penyampaian2} CaraPenyampaianInformasi3: {self.Penyampaian3} CaraPenyampaianInformasi4: {self.Penyampaian4} KontenInformasiYangDisampaikan: {self.Konten} BahasaYangDigunakanDalamPenyampaianInformasi: {self.Bahasa} PenguasaanMateri: {self.Penguasaan} MenjawabPertanyaan: {self.Menjawab} PenggunaanMediaPendukung: {self.Media} MenggunakanWaktuDenganEfektifDanEfisien: {self.Waktu}"
+
+class ResponseKuesioner(models.Model):
+    takes = models.ForeignKey(Takes, on_delete=models.CASCADE)
+    Kuesioner1 = models.IntegerField()
+    Kuesioner2 = models.IntegerField()
+    Kuesioner3 = models.IntegerField()
+    Kuesioner4 = models.IntegerField()
+    Kuesioner5 = models.IntegerField()
+    Kuesioner6 = models.IntegerField()
+    Kuesioner7 = models.IntegerField()
+    Kuesioner8 = models.IntegerField()
+    Kuesioner9 = models.IntegerField()
+    Kuesioner10 = models.IntegerField()
+    Kuesioner11 = models.IntegerField()
+    Kuesioner12 = models.IntegerField()
+
+    def __str__(self):
+        return f"Mahasiswa: {self.takes.student} Matkul: {self.takes.section.course_id} \
+        Informasi Yang Cukup: {self.Kuesioner1} \
+        Pelaksanaan: {self.Kuesioner2} \
+        Menguasai: {self.Kuesioner3} \
+        Perkuliahan Terorganisir: {self.Kuesioner4} \
+        Komunikasi Dosen: {self.Kuesioner5} \
+        Kepedulian Dosen: {self.Kuesioner6} \
+        Keadilan Dosen: {self.Kuesioner7} \
+        Beban Kerja Sesuai: {self.Kuesioner8} \
+        Sarana Prasarana: {self.Kuesioner9} \
+        Fasilitas Pendukung: {self.Kuesioner10} \
+        Usaha: {self.Kuesioner11} \
+        PengalamanBelajar: {self.Kuesioner12} \
+        "
+
+
+class LOSuplemenSemester(models.Model):
+    student = models.ForeignKey(Student, on_delete=models.CASCADE)
+    semester = models.IntegerField()
+    year = models.IntegerField()
+    lo_a = models.FloatField()
+    lo_b = models.FloatField()
+    lo_c = models.FloatField()
+    lo_d = models.FloatField()
+    lo_e = models.FloatField()
+    lo_f = models.FloatField()
+    lo_g = models.FloatField()
+
+class LOSuplemenCourse(models.Model):
+    student = models.ForeignKey(Student, on_delete=models.CASCADE)
+    section = models.ForeignKey(Section, on_delete=models.CASCADE)
+    lo_a = models.FloatField()
+    lo_b = models.FloatField()
+    lo_c = models.FloatField()
+    lo_d = models.FloatField()
+    lo_e = models.FloatField()
+    lo_f = models.FloatField()
+    lo_g = models.FloatField()
+
+class CourseAssessmentScore(models.Model):
+    semester = models.IntegerField()
+    year = models.IntegerField()
+    course = models.ForeignKey(Course, on_delete=models.CASCADE)
+
+    section_count = models.IntegerField(default = 1)
+    section_count_str = models.CharField(max_length = 9)
+
+    courseOutcomeScore = models.FloatField()
+    kuesionerScore = models.FloatField()
+    portfolioScore = models.FloatField(default = 4.0)
+
+    finalScore = models.FloatField()
+
+    def setCourseAssessment(self, semester, year, course_id, section_count, courseOutcomeScore, kuesionerScore, finalScore):
+        courseAssessment = CourseAssessmentScore.objects.filter(semester = semester, year = year, course__course_id = course_id)
+
+        if(len(courseAssessment) != 0):
+            courseAssessment[0].courseOutcomeScore = courseOutcomeScore
+            courseAssessment[0].kuesionerScore = kuesionerScore
+            courseAssessment[0].finalScore = finalScore
+            courseAssessment[0].section_count = section_count
+            courseAssessment[0].section_count_str = self.sectionCountView(self, section_count)
+            
+            courseAssessment[0].save()
+            return courseAssessment[0]
+        else:
+            course = Course.objects.get(course_id = course_id)
+            courseAssessmentScore = CourseAssessmentScore.objects.create(semester = semester, year = year, 
+                course = course, section_count = section_count, section_count_str = self.sectionCountView(self, section_count),
+                courseOutcomeScore = courseOutcomeScore, kuesionerScore = kuesionerScore, finalScore = finalScore)
+            return courseAssessmentScore
+
+    def sectionCountView(self, count):
+        if(count == 1):
+            return "K1"
+        else:
+            return "K1 - K" + str(count)
